requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.43.0
pinecone>=5.0.0
anthropic>=0.34.0
urllib3<2
streamlit>=1.35.0
